Add rounded box background to QR view
Use a white box with rounded corners as background for the QR / barcode
area, and put both the code and the alt text into it. This gives it a
nicer look by being more in line with the general rounded corners, makes
the code appear less "stamped" onto the pass, and improves alt text
readability depending on the pass background color.

This also closer how Apple Wallet shows the barcode block.