Restore full strip width (#323)
With the changes made in #310 the strip is now not expanded to the full
pass width anymore if the underlying image is smaller than the pass
width.

Restore the fillMaxWidth() modifier used on strip images, while keeping
the height / width modifier removed. This will make portait strip images
use the full width, resulting in a rather tall pass. In the case of DHL
it also addresses the image still being to small in #322.